backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
google-genai==0.3.0
requests==2.32.5
imageio-ffmpeg==0.6.0
pydub==0.25.1
SpeechRecognition==3.15.1
gTTS==2.5.4
gunicorn
